Refactor login, nav -menu
