+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "usereducer-end-game",
+  "name": "nftbaazar",
   "version": "1.0.0",
   "description": "",
   "keywords": [],
