+++ public/index.html
@@ -5,10 +5,9 @@
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
   <meta name="theme-color" content="#000000">
-  <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
-  <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
-  <link rel="stylesheet" href="https://nftbaazar.netlify.app/style.css">
+  <link rel="stylesheet" href="https://gitchui.netlify.app/style.css">
   <title>NFTbaazar</title>
+  <link rel="shortcut icon" href="../src/resources/title.gif">
 </head>
 
 <body>
