+++ src/Pages/Login/Login.css
@@ -1,0 +1,13 @@
+.form {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.login {
+    width: 100vw;
+    height: 80vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
