--- src/Pages/Logout.js
@@ -1,17 +1,0 @@
-import { useAuth } from "../Contexts";
-
-export function Logout() {
-  const { logoutUser } = useAuth();
-
-  const logoutHandler = () => {
-    logoutUser();
-  };
-
-  return (
-    <>
-      <button className="btn btn-error" onClick={logoutHandler}>
-        Logout :(
-      </button>
-    </>
-  );
-}
