+++ src/Pages/index.js
@@ -1,9 +1,8 @@
 export { Cart } from "./Cart";
 export { ProductListing } from "./ProductListing/ProductListing";
 export { WishList } from "./WishList";
-export { Login } from "./Login";
+export { Login } from "./Login/Login";
 export { Signup } from "./Signup";
-export { Logout } from "./Logout";
 export { ProductDetails } from "./ProductDetails/ProductDetails";
 export { TestAPI } from "./TestAPI";
 export { Home } from "./Home/Home";
